♻️ [Refactor]: props interface 파일 내부로 이동
