+++ src/app/(auth)/study-room/components/manage-applicants/InfoModal.tsx
@@ -8,9 +8,8 @@ import Modal, {
 import Profile from "@/shared/components/atoms/Profile";
 import Typography from "@/shared/components/atoms/Typography";
 
-import { StudyManageCardInterface } from "@/shared/types/api/study-room";
-
 import profileImg from "@/asset/images/profile_example.jpeg";
+import { StudyManageCardInterface } from "@/features/study-room/types/study-room.type";
 
 interface CurriculumModalProps extends ModalProps {
   onClose: ModalOnClose;
