+++ src/app/(auth)/study-room/components/manage-overview/ManageOverviewCard.tsx
@@ -3,7 +3,6 @@ import cn from "@/shared/utils/cn";
 import { useRouter, useSearchParams } from "next/navigation";
 import { useEffect, useState } from "react";
 
-import { ManageOverviewCardProps } from "@/features/study-room/types/study-room.type";
 import Card from "@/shared/components/atoms/Card";
 import Typography from "@/shared/components/atoms/Typography";
 import ActionConfirmModal from "@/shared/components/molecules/ActionConfirmModal";
@@ -15,6 +14,12 @@ const TABS = [
   { id: 1, label: "규칙" },
   { id: 2, label: "혜택" },
 ];
+
+interface ManageOverviewCardProps {
+  rules: string[];
+  benefits: string[];
+  studyId: string;
+}
 
 const ManageOverviewCard = ({
   rules,
